Phytoplankon data from the chlorpyrifos-nutrient (CPF-NUT) experiment from van den Brink & ter Braak (1998)

The dataframe CPTNUT_phytoplankton contains the design and phytoplankton data from the microcosm example of PRC from van den Brink & ter Braak (1998). In this experiment from van Donk et al. (1995), there are three treatments: control (con), nutrients (nut) and nutrients with chlorpyrifos application (cpf-nut), each with four cosms (replications) sampled 13 times (twice before nutrients where added and six times before the additional application of chorpyrifos). The phytoplankton data are counts and there are 26 species in total.

Format

A data frame with 156 rows and 31 columns:

samples character identifier for the sample.

Time factor coding for the week of sampling with respect to first application of nutrients.

Treatment treatment factor as in van den Brink & ter Braak (1998). The levels are con, nut and cpf-nut for the four control cosms, the four cosms that receive nutrients after the first two sampling points, and the four cosms to which both nutrients and chorpyrifos are applied, respectively. Note that the cosms of the nutrient-chlorpyrifos treatment are coded herein as nut in the weeks before the first chlorpyrifos application, as they are treated experiment-wise similarly in these weeks. But, note that the samples from con and nut are coded differently on pre-treatment times (although they are similar experiment-wise), so as to see the pre-treatment variability. See also the factor con_nut_cpf.

Plot numeric identifier for the microcosm.

con_nut_cpf treatment factor. Before application of nutrients or chlorpyrifos, the samples are all treated similarly, but are coded differently (as often in PRC).

cell counts of 25 different phytoplankton taxa in 1L, and density for the 26th species, Volvox. For details, see van den Donk et al. (1995).

Source

van Donk et al. (1995), van den Brink & ter Braak (1998)

References

van den Brink, PJ & ter Braak, CJF (1998). Multivariate analysis of stress on experimental ecosystems by principal response curves and similarity analysis. Aquatic ecology, 32, 163-178. doi:10.1023/A:1009944004756

van Donk E, Prins H, Voogd HM, Crum SJH and Brock TCM (1995) Effects of nutrient loading and insecticide application on the ecology of Elodea – dominated freshwater microcosms. I. Responses of plankton and zooplanktivorous insects. Arch Hydrobiol 133: 417–439. doi:10.1127/archiv-hydrobiol/133/1995/417