smooth PRC diagram

Creates a smooth PRC diagram with treatment curves and a vertical species-loading plot.

plotsmoothPRC(
  object,
  axis = 1,
  mod_prc = NULL,
  with_classical_lines = TRUE,
  with_unconstrained_scores = TRUE,
  flip = FALSE,
  widths = c(4, 1),
  title = NULL,
  left = "Treatment curves",
  right = "taxon scores",
  threshold = 14,
  y_lab_interval = 0.5,
  speciesname = NULL,
  selectname = "Fratio",
  sample_times_only = FALSE,
  verbose = TRUE
)

Arguments

object

a result of smoothPRC.

axis

the number of the axis to plot

mod_prc

a results of doPRC, i.e. a classical PRC with the same data and design. Default NULL, in which the classical PRC is computed using doPRC and the data in object.

with_classical_lines

logical to add or suppress classical PRC lines. Default TRUE, i.e. with classical PRC lines.

with_unconstrained_scores

logical to add or suppress the unconstrained scores in the smooth PRC showing the scatter of the individual samples (cosms). Default TRUE, i.e. with individual points.

flip

logical. Should the axis be reversed? Default FALSE. Can be numeric with -1 meaning: reverse, and 1 meaning do not reverse.

widths

relative width of treatment PRC plot and the loading plot (see grid.arrange).

title

overall title.

left

left axis text (see arrangeGrob).

right

right axis text (see arrangeGrob).

threshold

species with criterion (specified by selectname) higher than the threshold are displayed. Default: 7 (which is the threshold F-ratio for testing a single regression coefficient at p = 0.01 with 60 df for the error in a multiple regression of each single species onto the condition and the ordination axis). If selectname is not in species_scores, the threshold is divided by 14, so that the default is 0.5.

y_lab_interval

interval of the y-axis ticks. A tick at no effect (0) is always included; default: 0.5.

speciesname

name of the variable containing the species names (default NULL uses row names).

selectname

name of the column or variable containing the criterion for the selection of species to be displayed. Default: "Fratio1"; if selectname is not found in species_scores, set to scoresname.

sample_times_only

logical to PRClines based on the sampled times only, i.e. without interpolation. Default FALSE.

verbose

logical for printing the number of species with names out of the total number (default: TRUE).

Details

sample_times_only = TRUE can be useful for comparison with classical PRC. It avoids interpolation and may work when the default fails to produce a plot.